src/requirements.txt
blinker==1.9.0
certifi==2025.4.26
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
dnspython==2.7.0
Flask==3.1.0
#ganabosques-orm==0.1.0
git+https://github.com/CIAT-DAPA/ganabosques_orm
gsconfig-py3==1.0.7
idna==3.10
Flask-Login==0.6.3
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
mongoengine==0.29.1
pandas==2.2.3
pymongo==4.12.1
python-dotenv==1.1.0
requests==2.32.3
urllib3==2.4.0
Werkzeug==3.1.3
flask-wtf
wtforms
pandas
beautifulsoup4>=4.9.0
rasterio>=1.2.10
authlib



